refactor(doctor): extract state checks
